Add files via upload
clima is an online platform that allows users to obtain real-time and forecasted weather information for specific locations. The primary purpose of this website is to provide users with accurate and up-to-date details about current weather conditions, temperature, humidity, wind speed, and other relevant meteorological data for a chosen city or region.